requirements.txt
pandas
matplotlib
seaborn
scikit-learn
scikit-image
pytorch-ignite
pycocotools
grad-cam
pynvml